+++ docs/changelog.rst
@@ -55,7 +55,7 @@ Additions to the :doc:`supporting modules <supporting-modules>`.
     * :py:func:`khoros.utils.tests.test_version.test_latest_version`
 * Added the following functions to the :py:mod:`khoros.utils.tests.test_communities` module:
     * :py:func:`khoros.utils.tests.test_communities.set_package_path`
-    * :py:func:`khoros.utils.tests.test_communities.test_community_title`
+    * :py:func:`khoros.utils.tests.test_communities.test_community_details`
 
 |